xfs: rmap btree requires more reserved free space
The rmap btree is allocated from the AGFL, which means we have to
ensure ENOSPC is reported to userspace before we run out of free space
in each AG. The last allocation in an AG can cause a full height rmap
btree split, and that means we have to reserve at least this many
blocks *in each AG* to be placed on the AGFL at ENOSPC.  Update the
various space calculation functiosn to handle this.

v2: Because the macros are now executing conditional code and are
called quite frequently, convert them to functions that initialise
varaibles in the struct xfs_mount, use the new variables everywhere
and document the calculations better.

If rmapbt is disabled, it is incorrect to require 1 extra AGFL block
for the rmapbt (due to the + 1); the entire clause needs to be gated
on the feature flag.

This causes serious problems if formatting a v4 filesystem because the
extra AGFL block causes the root inode not to be where xfs_repair
expects it.  In turn, xfs_repair reports major FS damage when
everything is fine.

A few macros got replaced with functions, so update mkfs and repair to
use the new symbol names.  Fix a duplicate #define too.

Signed-off-by: Dave Chinner <dchinner@redhat.com>
[darrick.wong@oracle.com: v2 changes]
Signed-off-by: Darrick J. Wong <darrick.wong@oracle.com>