xfs: introduce the extended reverse mapping btree
Introduce the extended reverse mapping btree, or "rmapxbt", to handle
the larger keysize required to uniquely identify rmap records in a
reflink filesystem.

The "rmapbt" applies to non-reflink filesystems and can identify a
record solely from the physical block number.  The smaller keysize
results in a higher fanout factor for the btree, which makes it more
space-efficient.

The "rmapxbt" applies to reflink filesystems and requires a larger
key consisting of the physical block number, the owner, and an offset
within that owner, to identify a specific record.  This large-key
tree will eventually connect to the btree interval query support.

Fix up the userland utilities to understand both trees.

Signed-off-by: Darrick J. Wong <darrick.wong@oracle.com>