libxfs: calculate max btree size based on minimum btree block fill ratios
Calculate the maximum possible size of the rmap and refcount btrees
based on minimally-full btree blocks.  This increases the per-AG
block reservations to handle the worst case btree size.  In the end,
we reserve about ~0.7% space for refcount and ~1.3% space for rmap
when those features are enabled, to (try to) avoid ENOSPC at runtime.

Signed-off-by: Darrick J. Wong <darrick.wong@oracle.com>