libxfs: calculate (and range check) each btree's level separately
The refcount and rmap btrees have different record and key sizes from
the regular allocation trees.  Therefore, it is incorrect to use the
maximum allocation btree height to check either btree.  Refactor the
code into a common btree function and per-btree wrappers, and use the
appropriate wrapper to check the refcount and rmap height.

(Usually the per-AG structure will be live, so we probably won't hit
this code.)

Signed-off-by: Darrick J. Wong <darrick.wong@oracle.com>