xfs: account for per-AG reservation in statfs f_blocks
Since the blocks reserved by the per-AG reservation mechanism are never
available to userspace, there's no point in reporting them via statfs.
Reduce the number of blocks reported by statfs so our space accounting
works the way it did in the old days -- f_blocks is the theoretical
upper bound on the amount of space that user programs could allocate,
and f_blocks is the current maximum.

This eliminates the regression where you format a 100T XFS and df
reports 2T are already "used".  Now it reports that you have a 98T
filesystem.

(Dave's thinp rfc might very well fix this whole problem; this is
purely a bandaid to shut down the complaints.)

Signed-off-by: Darrick J. Wong <darrick.wong@oracle.com>